Rename chart.js to Chart.js
